+++ History.md
@@ -2,6 +2,7 @@ unreleased
 ==========
 
   * Move to `jshttp` orgainzation
+  * Prevent `arguments`-related de-opt
 
 0.0.0 / 2014-05-13
 ==================